requirements.txt
Flask==1.1.2
numpy==1.18.4
tensorflow-estimator==2.2.0
tensorflow-cpu==2.2.0
